requirements.txt
asgiref==3.8.1
Django==5.0.7
django-jalali==7.0.0
django-jalali-date==1.1.3
jalali_core==1.0.0
jdatetime==5.0.0
sqlparse==0.5.1
tzdata==2024.1
